[DDP] Generalize activation checkpoint tests
enable these tests to run for all dist backends not just nccl.

Differential Revision: [D34281684](https://our.internmc.facebook.com/intern/diff/D34281684/)

[ghstack-poisoned]
