+++ test/distributed/test_c10d_nccl.py
@@ -949,7 +949,7 @@ class ProcessGroupNCCLTest(MultiProcessTestCase):
 
 
 class DistributedDataParallelTest(
-    test_c10d_common.AbstractDistributedDataParallelTest, MultiProcessTestCase
+    test_c10d_common.CommonDistributedDataParallelTest, MultiProcessTestCase
 ):
     def setUp(self):
         super(DistributedDataParallelTest, self).setUp()
@@ -957,6 +957,10 @@ class DistributedDataParallelTest(
         # that use NCCL_BLOCKING_WAIT will test it as expected.
         os.environ["NCCL_ASYNC_ERROR_HANDLING"] = "1"
         self._spawn_processes()
+
+    def _get_process_group(self):
+        store = self._get_store()
+        return c10d.ProcessGroupNCCL(store, self.rank, self.world_size)
 
     def _test_nccl_backend(
         self, devices, device_ids, multi_device=False, gradient_as_bucket_view=False
@@ -2216,349 +2220,6 @@ class DistributedDataParallelTest(
                         ),
                     )
 
-    # A list of tests for ddp with activation checkpointing
-    # when gradient_as_bucket_view=True, False.
-    # Most of the tests are referred to
-    # https://github.com/facebookresearch/fairscale/blob/main/tests/nn/pipe/test_checkpoint_ddp.py
-    class CheckpointOnceModule(nn.Module):
-        """
-        Runs checkpoint for a single layer in the model.
-        """
-        def __init__(self, use_reentrant=True):
-            super().__init__()
-            self.l1 = nn.Linear(20, 20)
-            self.l2 = nn.Linear(20, 20)
-            self.use_reentrant = use_reentrant
-
-        def forward(self, inp):
-            x = self.l1(inp)
-            x = checkpoint(self.l2, x, use_reentrant=self.use_reentrant)
-            return x
-
-    class CheckpointTwiceModule(CheckpointOnceModule):
-        """
-        Runs checkpoint for the same layer twice in a model. This simulates use
-        cases such as pipeline parallel where the same layer can be checkpointed
-        more than one time.
-        """
-        def __init__(self, use_reentrant=True):
-            super().__init__(use_reentrant=use_reentrant)
-
-        def forward(self, inp):
-            x = self.l1(inp)
-            x = checkpoint(self.l2, x, use_reentrant=self.use_reentrant)
-            x = checkpoint(self.l2, x, use_reentrant=self.use_reentrant)
-            return x
-
-    class CheckpointTwiceModuleWeightSharing(CheckpointTwiceModule):
-        """
-        Similar to CheckpointTwiceModule but the weights are shared.
-        """
-        def __init__(self, use_reentrant=True):
-            super().__init__(use_reentrant=use_reentrant)
-
-        def forward(self, inp):
-            x = self.l1(inp)
-            x = checkpoint(self.l2, x, use_reentrant=self.use_reentrant)
-            x = checkpoint(self.l2, x, use_reentrant=self.use_reentrant)
-            return x
-
-
-    class DynamicCheckpointTwiceModule(CheckpointTwiceModule):
-        def __init__(self, use_reentrant=True):
-            super().__init__(use_reentrant=use_reentrant)
-            self.count = 0
-
-        def forward(self, inp):
-            if self.count % 2:
-                x = checkpoint(self.l1, inp, use_reentrant=self.use_reentrant)
-            else:
-                x = checkpoint(self.l2, inp, use_reentrant=self.use_reentrant)
-
-            self.count += 1
-            return x
-
-    class DynamicCheckpointTwiceModuleWeightSharing(DynamicCheckpointTwiceModule):
-        def __init__(self, use_reentrant=True):
-            super().__init__(use_reentrant=use_reentrant)
-            self.l1.weight = self.l2.weight
-
-
-    def _prepare_dummy_data(self):
-        ddp_bs = 16
-        bs = ddp_bs * self.world_size
-        input = torch.rand((bs, 20), device="cuda", requires_grad=True)
-        target = torch.randn((bs, 20), device="cuda")
-        offset = self.rank * ddp_bs
-        ddp_input = input[offset : offset + ddp_bs]
-        ddp_target = target[offset : offset + ddp_bs]
-        return input, ddp_input, target, ddp_target
-
-    def _train_model(self, model, input_var, target, loss, run_checkpoint=False, use_reentrant=True):
-        model.train()
-        if run_checkpoint:
-            output = checkpoint(model, input_var, use_reentrant=use_reentrant)
-        else:
-            output = model(input_var)
-        l = loss(output, target)
-        l.backward()
-
-    def _test_ddp_checkpointing(
-        self,
-        input_model,
-        process_group,
-        use_bucket_view,
-        find_unused_parameters=False,
-        static_graph=False,
-        run_checkpoint=False,
-        use_reentrant=True,
-        allow_none_grads=False,
-    ):
-        # to reproduce the same training results
-        torch.cuda.set_device(self.rank)
-        torch.manual_seed(31415)
-        model = copy.deepcopy(input_model).cuda()
-        ddp_model = copy.deepcopy(input_model).cuda()
-        ddp_model = nn.parallel.DistributedDataParallel(
-            ddp_model,
-            bucket_cap_mb=1,
-            gradient_as_bucket_view=use_bucket_view,
-            device_ids=[self.rank],
-            process_group=process_group,
-            find_unused_parameters=find_unused_parameters,
-            static_graph=static_graph,
-        )
-        self.assertEqual(
-            ddp_model._get_ddp_logging_data().get("static_graph", 0), static_graph
-        )
-        input, ddp_input, target, ddp_target = self._prepare_dummy_data()
-        loss = nn.MSELoss()
-        n_iters = 5
-        for i in range(n_iters):
-            model.zero_grad(set_to_none=False)
-            ddp_model.zero_grad(set_to_none=False)
-            self._train_model(model, input, target, loss, run_checkpoint=run_checkpoint, use_reentrant=use_reentrant)
-            self._train_model(
-                ddp_model, ddp_input, ddp_target, loss, run_checkpoint=run_checkpoint, use_reentrant=use_reentrant
-            )
-            for i, j in zip(model.parameters(), ddp_model.parameters()):
-                if not allow_none_grads:
-                    self.assertTrue(i.grad is not None)
-                    self.assertTrue(j.grad is not None)
-                self.assertEqual(i.grad, j.grad, rtol=1.3e-06, atol=5e-5)
-
-    @requires_nccl()
-    @skip_if_lt_x_gpu(2)
-    @parametrize("use_reentrant", [True, False])
-    def test_ddp_checkpointing_once(self, use_reentrant):
-        """
-        DDP works as expected when layer is checkpointed only once.
-        """
-        store = c10d.FileStore(self.file_name, self.world_size)
-        process_group = c10d.ProcessGroupNCCL(store, self.rank, self.world_size)
-        for use_bucket_view, static_graph in product((False, True), (False, True)):
-            self._test_ddp_checkpointing(
-                self.CheckpointOnceModule(use_reentrant=use_reentrant),
-                process_group=process_group,
-                use_bucket_view=use_bucket_view,
-                static_graph=static_graph,
-            )
-            if static_graph:
-                # find_unused_parameters does not make a difference, since it is
-                # ignored for static graph.
-                self._test_ddp_checkpointing(
-                    self.CheckpointOnceModule(),
-                    process_group=process_group,
-                    use_bucket_view=use_bucket_view,
-                    static_graph=static_graph,
-                    find_unused_parameters=True,
-                )
-
-    @requires_nccl()
-    @skip_if_lt_x_gpu(2)
-    @parametrize("use_reentrant", [True, False])
-    def test_ddp_checkpointing_unused_params(self, use_reentrant):
-        """
-        With reentrant autograd checkpointing impl, DDP will fail when there are
-        unused params in the model and no static graph training. With
-        non-reentrant checkpointing implementation, this works as expected.
-        """
-        store = c10d.FileStore(self.file_name, self.world_size)
-        process_group = c10d.ProcessGroupNCCL(store, self.rank, self.world_size)
-        for use_bucket_view in (True, False):
-            err_ctx = (
-                suppress() if not use_reentrant else
-                self.assertRaisesRegex(
-                    RuntimeError,
-                    "Expected to mark a variable ready only once."
-                )
-            )
-            with err_ctx:
-                model = self._test_ddp_checkpointing(
-                    self.CheckpointOnceModule(use_reentrant=use_reentrant),
-                    process_group=process_group,
-                    use_bucket_view=use_bucket_view,
-                    find_unused_parameters=True,
-                )
-            # test passes when static_graph is true
-            model = self._test_ddp_checkpointing(
-                self.CheckpointOnceModule(use_reentrant=use_reentrant),
-                process_group=process_group,
-                use_bucket_view=use_bucket_view,
-                find_unused_parameters=True,
-                static_graph=True,
-            )
-
-    @requires_nccl()
-    @skip_if_lt_x_gpu(2)
-    @parametrize("use_reentrant", [True, False])
-    def test_ddp_checkpointing_twice(self, use_reentrant):
-        """
-        Checkpoitning twice fails for non-static graph with reentrant checkpoint
-        implementation, succeeds with non-reentrant checkpoint implementation.
-        """
-        store = c10d.FileStore(self.file_name, self.world_size)
-        process_group = c10d.ProcessGroupNCCL(store, self.rank, self.world_size)
-        for use_bucket_view in (True, False):
-            err_ctx = (
-                suppress() if not use_reentrant else
-                self.assertRaisesRegex(
-                    RuntimeError,
-                    "Expected to mark a variable ready only once."
-                )
-            )
-            with err_ctx:
-                model = self._test_ddp_checkpointing(
-                    self.CheckpointTwiceModule(use_reentrant=use_reentrant),
-                    process_group=process_group,
-                    use_bucket_view=use_bucket_view,
-                    static_graph=False,
-                )
-
-            with err_ctx:
-                model = self._test_ddp_checkpointing(
-                    self.CheckpointTwiceModule(use_reentrant=use_reentrant),
-                    process_group=process_group,
-                    use_bucket_view=use_bucket_view,
-                    static_graph=False,
-                    find_unused_parameters=True,
-                )
-
-    @requires_nccl()
-    @skip_if_lt_x_gpu(2)
-    @parametrize("use_reentrant", [True, False])
-    def test_ddp_checkpointing_twice_static_graph(self, use_reentrant):
-        """
-        Regardless of reentrant or non-reentrant checkpointing impl,
-        checkpointing twice works with static graph enabled.
-        """
-        store = c10d.FileStore(self.file_name, self.world_size)
-        process_group = c10d.ProcessGroupNCCL(store, self.rank, self.world_size)
-        for use_bucket_view in (True, False):
-            # Test passes when static_graph=True.
-            model = self._test_ddp_checkpointing(
-                self.CheckpointTwiceModule(use_reentrant=use_reentrant),
-                process_group=process_group,
-                use_bucket_view=use_bucket_view,
-                static_graph=True,
-            )
-
-    @requires_nccl()
-    @skip_if_lt_x_gpu(2)
-    def test_ddp_checkpointing_dynamic_module(self):
-        """
-        Dynamic module can be checkpointed, multiple times, with non-reentrant
-        checkpointing implementation.
-        """
-        store = c10d.FileStore(self.file_name, self.world_size)
-        process_group = c10d.ProcessGroupNCCL(store, self.rank, self.world_size)
-        for use_bucket_view in (True, False):
-            model = self._test_ddp_checkpointing(
-                self.DynamicCheckpointTwiceModule(use_reentrant=False),
-                process_group=process_group,
-                use_bucket_view=use_bucket_view,
-                static_graph=False,
-                find_unused_parameters=True,
-                # Grads can be none sometimes due to dynamic module not using
-                # all params.
-                allow_none_grads=True
-            )
-
-    @requires_nccl()
-    @skip_if_lt_x_gpu(2)
-    def test_ddp_checkpointing_dynamic_weight_sharing(self):
-        """
-        Dynamic module can be checkpointed multiple times with weight sharing
-        using non-reentrant checkpointing implementation.
-        """
-        store = c10d.FileStore(self.file_name, self.world_size)
-        process_group = c10d.ProcessGroupNCCL(store, self.rank, self.world_size)
-        for use_bucket_view in (True, False):
-            model = self._test_ddp_checkpointing(
-                self.DynamicCheckpointTwiceModuleWeightSharing(use_reentrant=False),
-                process_group=process_group,
-                use_bucket_view=use_bucket_view,
-                static_graph=False,
-                find_unused_parameters=True,
-                # Grads can be none sometimes due to dynamic module not using
-                # all params.
-                allow_none_grads=True
-            )
-
-    # DDP works as expected if there is weight sharing among layers
-    @requires_nccl()
-    @skip_if_lt_x_gpu(2)
-    @parametrize("use_reentrant", [True, False])
-    def test_ddp_checkpointing_weight_sharing(self, use_reentrant):
-        """
-        Test that checkpointing with weight sharing works.
-        """
-        store = c10d.FileStore(self.file_name, self.world_size)
-        process_group = c10d.ProcessGroupNCCL(store, self.rank, self.world_size)
-        torch.cuda.set_device(self.rank)
-        for use_bucket_view, static_graph in product((False, True), (False, True)):
-            torch.manual_seed(31415)
-            l1 = nn.Linear(20, 20)
-            l2 = nn.Linear(20, 20)
-            l1.weight = l2.weight
-            model = nn.Sequential(l1, l2)
-            # TODO: non-reentrant based checkpointing of DDP module with
-            # static_graph runs into the below issue, see
-            # https://github.com/pytorch/pytorch/issues/70865 and
-            # https://github.com/pytorch/pytorch/issues/58111 for details.
-            err_ctx = (
-                self.assertRaisesRegex(
-                    RuntimeError,
-                    "Your training graph has changed in this iteration"
-                ) if static_graph and not use_reentrant else suppress()
-            )
-            with err_ctx:
-                self._test_ddp_checkpointing(
-                    model,
-                    process_group=process_group,
-                    use_bucket_view=use_bucket_view,
-                    static_graph=static_graph,
-                    run_checkpoint=True,
-                    use_reentrant=use_reentrant,
-                )
-
-    @requires_nccl()
-    @skip_if_lt_x_gpu(2)
-    def test_ddp_checkpointing_twice_weight_sharing(self):
-        """
-        Checkpointing should work with static graph in the case of checkpointing
-        same layer twice and having weights shared acrosss layers.
-        """
-        store = c10d.FileStore(self.file_name, self.world_size)
-        process_group = c10d.ProcessGroupNCCL(store, self.rank, self.world_size)
-        torch.cuda.set_device(self.rank)
-        for use_bucket_view in (True, False):
-            model = self._test_ddp_checkpointing(
-                self.CheckpointTwiceModuleWeightSharing(),
-                process_group=process_group,
-                use_bucket_view=use_bucket_view,
-                static_graph=True,
-            )
 
 
 class NcclErrorHandlingTest(MultiProcessTestCase):
@@ -3053,7 +2714,7 @@ class CommTest(test_c10d_common.AbstractCommTest, MultiProcessTestCase):
     def test_nccl_warn_not_in_group_debug_off(self):
         self._test_warn_not_in_group(backend="nccl")
 
-instantiate_parametrized_tests(DistributedDataParallelTest)
+# instantiate_parametrized_tests(DistributedDataParallelTest)
 
 if __name__ == "__main__":
     assert (
